move str opimpls into their own file
